requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
instagrapi==2.1.5
pillow==11.2.1
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
